feat: Course, Apply, Member - Course 리파지토리 Custom 작성중 - Apply 리파지토리 작성 - Member 서비스, dto 추가
test: Course, Apply 테스트 코드 작성
- 리파지토리 테스트 완료
- 서비스 테스트

chore: 오타 수정
- Auditing 어노테이션 추가
- 오타 수정(DogRepository, Apply)

+++ src/main/java/com/bbodeum/dog/repository/DogRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.bbodeum.dog.entity.Dog;
 import com.bbodeum.member.entity.Member;
 
-public interface DogRepository extends CrudRepository<Dog, Integer> {
+public interface DogRepository extends CrudRepository<Dog, Long> {
 	public List<Dog> findByMember(Member m);
 	public List<Dog> findByMemberAndDogName(Member m, String dogName);
